test: split wallet lifecycle non-repair tests
Changed:
Split the remaining wallet lifecycle tests into expiration, grants, snapshots, and usage modules without changing test behavior.

Benefit:
The billing wallet lifecycle suite is easier to navigate and maintain after the repair tests were separated.
